Count logic for down method

+++ src/main/kotlin/com/improve_future/harmonica/core/Harmonica.kt
@@ -24,9 +24,9 @@ open class Harmonica(private val packageName: String, private val migrationTable
             for (clazz in allClasses.sortedBy { it.simpleName }) {
                 val migrationVersion: String = versionService.pickUpVersionFromClassName(clazz.simpleName)
 
-                if (applyCount !== null && migrationCount >= applyCount) {
-                    break;
-                }
+                if (applyCount !== null && migrationCount >= applyCount)
+                    break
+
                 if (versionService.isVersionMigrated(connection, migrationVersion)) continue
                 migrationCount++
 
@@ -50,10 +50,10 @@ open class Harmonica(private val packageName: String, private val migrationTable
     /**
      * Down latest migrations
      */
-    fun down(db: Database, applyCount: Int? = null) {
+    fun down(db: Database, applyCount: Int = 1) {
         val connection = ExposedConnection(db = db)
         try {
-            val migrationVersion = versionService.findCurrentMigrationVersion(connection)
+            val migrationVersionList = versionService.findListMigrationVersion(connection, applyCount)
 
             val reflections = Reflections(packageName)
             val allClasses = reflections.getSubTypesOf(AbstractMigration::class.java)
@@ -61,26 +61,29 @@ open class Harmonica(private val packageName: String, private val migrationTable
             if (allClasses.isEmpty())
                 return
 
-            val fileCandidateList: List<Class<out AbstractMigration>> =
-                allClasses.filter { it.simpleName.startsWith(migrationVersion) }
+            for (migrationVersion in migrationVersionList) {
+                val fileCandidateList: List<Class<out AbstractMigration>> =
+                    allClasses.filter { it.simpleName.contains(migrationVersion) }
 
-            if (1 < fileCandidateList.size)
-                throw Error("More then one files exist for migration $migrationVersion")
+                if (fileCandidateList.isEmpty())
+                    throw Error("Not found $migrationVersion")
 
-            val clazz = allClasses.first()
+                if (1 < fileCandidateList.size)
+                    throw Error("More then one files exist for migration $migrationVersion")
 
-            println("== [Start] Migrate down $migrationVersion ==")
-            connection.transaction {
-                val kClass = clazz.kotlin
-                val migration: AbstractMigration = kClass.createInstance()
+                val clazz = fileCandidateList.first()
 
-                migration.connection = connection
-                migration.down()
-                versionService.removeVersion(connection, migrationVersion)
+                println("== [Start] Migrate down $migrationVersion ==")
+                connection.transaction {
+                    val kClass = clazz.kotlin
+                    val migration: AbstractMigration = kClass.createInstance()
+
+                    migration.connection = connection
+                    migration.down()
+                    versionService.removeVersion(connection, migrationVersion)
+                }
+                println("== [End] Migrate down $migrationVersion ==")
             }
-
-            println("== [End] Migrate down $migrationVersion ==")
-
         } catch (e: Exception) {
             throw e
         } finally {
